Created camlQuery.js
Added a useful amd_module to this project that abstracts much of the caml query details from the developer.  Helps to avoid using additional third party libraries and is very simple to understand if you are used to writing caml queries.  Not good if you rather use something like pnp or SPServices because caml queries can be a bit of a pain sometimes.